pyproject.toml
[project]
name = "rollercoaster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
	"grpcio>=1.76.0",
	"grpcio-tools>=1.76.0",
]

[dependency-groups]
dev = [
	"ruff>=0.14.4",
]

[tool.ruff.lint]
select = [
	# pycodestyle
	"E",
	# Pyflakes
	"F",
	# pyupgrade
	"UP",
	# flake8-bugbear
	"B",
	# flake8-simplify
	"SIM",
	# isort
	"I",
]
ignore = ["E501"]
[tool.ruff.format]
quote-style = "single"
indent-style = "tab"
docstring-code-format = true

[tool.basedpyright]
exclude = ["proto/"]